requirements.txt
# General dependencies
numpy>=1.20.0
python-dotenv>=0.15.0
requests>=2.25.1
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.3  # PostgreSQL driver

# Testing and linting
flake8>=3.9.0
mypy>=0.812
black>=21.5b2
isort>=5.8.0

# Optional libraries for advanced features
tqdm>=4.60.0  # Progress bars